+++ exonum/src/blockchain/tests.rs
@@ -627,7 +627,7 @@ fn test_dispatcher_start_service_good() {
     // Tests start service for the good service.
     let snapshot = blockchain.snapshot();
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-1".to_owned()));
     execute_transaction(
         &mut blockchain,
@@ -635,7 +635,7 @@ fn test_dispatcher_start_service_good() {
     );
     let snapshot = blockchain.snapshot();
     assert!(DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-1".to_owned()));
     assert_eq!(snapshot.get_entry(IDX_NAME).get(), Some(1_u64));
 }
@@ -649,7 +649,7 @@ fn test_dispatcher_start_service_rollback() {
     // Tests that a service with an unregistered artifact will not be started.
     let snapshot = blockchain.snapshot();
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-24".to_owned()));
     execute_transaction(
         &mut blockchain,
@@ -657,13 +657,13 @@ fn test_dispatcher_start_service_rollback() {
     );
     let snapshot = blockchain.snapshot();
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-24".to_owned()));
     assert!(!snapshot.get_entry::<_, u64>(IDX_NAME).exists());
 
     // Tests that a service with panic during the configure will not be started.
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-42".to_owned()));
     execute_transaction(
         &mut blockchain,
@@ -671,13 +671,13 @@ fn test_dispatcher_start_service_rollback() {
     );
     let snapshot = blockchain.snapshot();
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-42".to_owned()));
     assert!(!snapshot.get_entry::<_, u64>(IDX_NAME).exists());
 
     // Tests that a service with execution error during the configure will not be started.
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-18".to_owned()));
     execute_transaction(
         &mut blockchain,
@@ -685,7 +685,27 @@ fn test_dispatcher_start_service_rollback() {
     );
     let snapshot = blockchain.snapshot();
     assert!(!DispatcherSchema::new(&snapshot)
-        .service_instances()
+        .instances()
         .contains(&"good-service-18".to_owned()));
     assert!(!snapshot.get_entry::<_, u64>(IDX_NAME).exists());
+}
+
+/// Checks that `BlockchainMut::check_tx` discards transactions with incorrect
+/// instance IDs.
+#[test]
+fn test_check_tx() {
+    let keypair = crypto::gen_keypair();
+    let blockchain = create_blockchain(vec![InstanceCollection::new(TestDispatcherService)
+        .with_instance(TEST_SERVICE_ID, TEST_SERVICE_NAME, ())]);
+
+    let correct_tx = TestAdd { value: 1 }.sign(TEST_SERVICE_ID, keypair.0, &keypair.1);
+
+    assert_eq!(blockchain.check_tx(&correct_tx), Ok(()));
+
+    let incorrect_tx = TestAdd { value: 1 }.sign(TEST_SERVICE_ID + 1, keypair.0, &keypair.1);
+
+    assert_eq!(
+        blockchain.check_tx(&incorrect_tx),
+        Err(DispatcherError::IncorrectInstanceId.into())
+    );
 }