+++ CHANGELOG.md
@@ -46,11 +46,11 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
   - Removed obsolete `TestKit::blockchain_mut` method and `TestKit::blockchain`
   now returns value instead of reference.
+  
+- Dot symbol is not allowed in service names anymore. (#1558)
 
 - Services can now use `BlockchainData` and `SnapshotExt` types to access data
   from the blockchain in a more structured manner. (#1523)
-
-- Placeholder for changes due to dynamic services (#9999)
 
 #### exonum-merkledb
 
@@ -70,6 +70,13 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   corresponds to a separate index or a group of indexes. It is possible
   to derive a constructor for such schemas via `FromAccess` derive macro.
   (#1523, #1562)
+
+- New index name restrictions has been added. (#1558)
+
+  - Dot symbol is not allowed anymore in indexes with prefixed access.
+
+  - Index names starting from `__` and not containing a dot `.` are reserved and
+    used only for system indexes.
 
 #### exonum-proto
 
@@ -450,6 +457,8 @@ Key points:
 
 - #1555: Update MerkleDB docs
 
+- #1568: Make DispatcherSchema merkelized again [ECR-3810]
+
 </details>
 <!-- markdownlint-enable no-inline-html -->
 
@@ -462,6 +471,10 @@ Key points:
   They are intended to configure restart settings of the HTTP-server (#1536).
 
 - `exonum` now has a `python` library for implementing integration tests. (#1516)
+
+- `BlockchainMut` now has a `check_tx` method used to verify transactions before
+  adding them to the transactions pool. Transactions for which `check_tx` fails
+  are considered invalid and can't be included to the block. (#1579)
 
 #### exonum-merkledb
 
@@ -509,6 +522,10 @@ Key points:
   performance improvements for large lists. (#1455)
 
 - Proofs building mechanisms have been heavily refactored. (#1460)
+
+#### exonum-testkit
+
+- Configuration change example has been moved to `exonum-supervisor` crate. (#1582)
 
 #### exonum-build